refactor: extract tracing configuration
The main module can be stripped of all the functionality related to
tracing and kept simple. Extract all functions and types that are
required for configuring tracing into it's own module.
