+++ src/main.rs
@@ -23,8 +23,11 @@ mod reaction_roles;
 mod scheduler;
 /// A trait to define a job that needs to be executed regularly, for example checking for status updates daily.
 mod tasks;
+mod trace;
 mod utils;
 
+use crate::trace::setup_tracing;
+use crate::trace::ReloadHandle;
 use anyhow::Context as _;
 use poise::{Context as PoiseContext, Framework, FrameworkOptions, PrefixFrameworkOptions};
 use reaction_roles::{handle_reaction, populate_data_with_reaction_roles};
@@ -33,113 +36,16 @@ use serenity::{
     client::{Context as SerenityContext, FullEvent},
     model::gateway::GatewayIntents,
 };
-use tokio::sync::RwLock;
 use tracing::info;
-use tracing_subscriber::{
-    fmt,
-    layer::SubscriberExt,
-    reload::{self, Layer},
-    EnvFilter, Registry,
-};
 
-use std::{
-    collections::{HashMap, HashSet},
-    fs::File,
-    sync::Arc,
-};
+use std::collections::{HashMap, HashSet};
 
 pub type Error = Box<dyn std::error::Error + Send + Sync>;
 pub type Context<'a> = PoiseContext<'a, Data, Error>;
-pub type ReloadHandle = Arc<RwLock<reload::Handle<EnvFilter, Registry>>>;
 
 pub struct Data {
     pub reaction_roles: HashMap<ReactionType, RoleId>,
     pub log_reload_handle: ReloadHandle,
-}
-
-/// Environment variables that our tracing configuration relies on
-///
-/// # Fields
-///
-/// * env: String that decides in what context the application will be running on i.e "production" or "development". This allows us to filter out logs from `stdout` when in production. Possible TODO: Could be replaced to a boolean `is_dev` or something similar to be more constrained than a string.
-/// * enable_debug_libraries: Boolean flag that controls whether tracing will output logs from other crates used in the project. This is only needed for really serious bugs.
-struct TracingConfig {
-    env: String,
-    enable_debug_libraries: bool,
-}
-
-impl TracingConfig {
-    /// Encapsulate all the required env variables into a [`TracingConfig`]
-    fn load_tracing_config() -> Self {
-        Self {
-            env: std::env::var("AMD_RUST_ENV").unwrap_or("development".to_string()),
-            // Some Rust shenanigans to set the default value to a boolean false:
-            enable_debug_libraries: std::env::var("ENABLE_DEBUG_LIBRARIES")
-                .unwrap_or("false".to_string())
-                .parse()
-                .unwrap_or(false),
-        }
-    }
-}
-
-/// Return the appropriate String denoting the level and breadth of logs depending on the [`TracingConfig`] passed in.
-fn build_filter_string(config: &TracingConfig) -> String {
-    let crate_name = env!("CARGO_CRATE_NAME");
-
-    match (config.env.as_str(), config.enable_debug_libraries) {
-        ("production", true) => "info".to_string(),
-        ("production", false) => format!("{crate_name}=info"),
-
-        (_, true) => "trace".to_string(),
-        (_, false) => format!("{crate_name}=trace"),
-    }
-}
-
-/// Build a suitable subscriber based on the context of the environment (i.e production or development). The only difference in subscriber configuration is that in a production context, logs are only sent to `amd.log` and not to `stdout`. This is done on the assumption that when deployed in production, checking terminal logs is neither reliable nor convenient.
-///
-/// # Arguments
-///
-/// * env: A string that can be set to "production" in order to disable logging to `stdout` and when set to anything else, enable logging to `stdout`.
-/// * filter: The filter used to determine the log level for this subscriber.
-///
-/// Returns the initialized subscriber inside a [`Box`].
-fn build_subscriber<L>(
-    env: String,
-    filter: L,
-) -> anyhow::Result<Box<dyn tracing::Subscriber + Send + Sync>>
-where
-    L: tracing_subscriber::Layer<tracing_subscriber::Registry> + Send + Sync + 'static,
-{
-    let file_layer = fmt::layer()
-        .pretty()
-        .with_ansi(false)
-        .with_writer(File::create("amd.log").context("Failed to create log file")?);
-
-    if env != "production" {
-        Ok(Box::new(
-            tracing_subscriber::registry()
-                .with(filter)
-                .with(file_layer)
-                .with(fmt::layer().pretty().with_writer(std::io::stdout)),
-        ))
-    } else {
-        Ok(Box::new(
-            tracing_subscriber::registry().with(filter).with(file_layer),
-        ))
-    }
-}
-
-fn setup_tracing() -> anyhow::Result<ReloadHandle> {
-    let config = TracingConfig::load_tracing_config();
-    let filter_string = build_filter_string(&config);
-    let (filter, reload_handle) = Layer::new(EnvFilter::new(filter_string));
-
-    let boxed_subscriber: Box<dyn tracing::Subscriber + Send + Sync> =
-        build_subscriber(config.env, filter)?;
-    tracing::subscriber::set_global_default(boxed_subscriber)
-        .context("Failed to set subscriber")?;
-
-    Ok(Arc::new(RwLock::new(reload_handle)))
 }
 
 #[tokio::main]
